Show owner import confirmation in base overview fragment
