+++ app/src/main/res/layout/fragment_transactions.xml
@@ -7,20 +7,20 @@
     android:layout_height="match_parent"
     tools:background="@color/background">
 
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tx_tab_bar"
-        style="@style/TabLayout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/tx_content"
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tx_tab_bar" />
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tx_tab_bar"
+            style="@style/TabLayout" />
+
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/tx_content"
+            android:layout_width="match_parent"
+            android:layout_height="0dp" />
+
+    </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
